[Refactor] 디렉터리 이동
